add table container
